Use ThreadPoolExecutor
